code enhancement

+++ PoGoBot.py
@@ -2,42 +2,41 @@ import discord
 from discord.ext.commands import Bot
 from discord.ext import commands
 import asyncio
-import time
 import datetime
 import re
 import os
 from Raid import *
 from Channel import *
 from data import *
+from utils import *
 
 Client = discord.Client()
 client = commands.Bot(command_prefix = "?")
 
 cookieCompteur = 0
 cRaids = {}
-activeChannels = 0
 server = 0
 
 cAccueil = 0
-cRaidList = 0
 cDiscussion = 0
 cPokemon = 0
 cRaidAdd = 0
 
-def is_bot(m):
-    global msgRaid
-    return m.author.bot != True
-
-#gesionnaire des listes de RAid
+#gesionnaire de la liste de RAid
 async def addToListe(cRaid):
     """ajoute un raid à la liste des raids"""
     #variables externes
     global cRaidAdd
 
+    if not isinstance(cRaid, ChannelRaid):
+        print ('bite')
+        return 0
+
     await client.purge_from(cRaidAdd, check=is_bot)
     content = str("raid en cour sur <#%s>" %(cRaid.com.id))
     msg = await client.send_message(cRaidAdd, content=content, embed=cRaid.raid.embed())
     cRaid.listMsg = msg
+    return 1
 async def editListe(cRaid):
     """editer le message corresponant au raid selectionné"""
     msg = str("raid en cour sur <#%s>" %(cRaid.com.id))
@@ -48,10 +47,6 @@ async def removeFromListe(cRaid):
     global cRaidAdd
 
     await client.delete_message(cRaid.listMsg)
-
-
-
-
     return 0
 
 #gestionnaire des Raid channels du forum
@@ -60,20 +55,49 @@ async def removeCRaid(cRaid):
     renvoit 1 si reussi
     0 sinon"""
     if not isinstance(cRaid, ChannelRaid): return 0
-    await client.send_message(cRaid.com, "Attention dans 1 min je vais détruire ce salon, n'oubliez pas de vous dire au revoir !")
-    await asyncio.sleep(60)
+    #await client.send_message(cRaid.com, "Attention dans 1 min je vais détruire ce salon, n'oubliez pas de vous dire au revoir !")
     cId = cRaid.com.id
     await removeFromListe(cRaid)
     await client.delete_channel(client.get_channel(cId))
     return 1
+async def editCRaid(cRaid):
+    """edite les information du raid
+    renvoit 1 si reussi
+    0 sinon"""
+    if not isinstance(cRaid, ChannelRaid): return 0
+    await client.edit_message(cRaid.pinMsg, embed=cRaid.raid.embed())
+    await client.edit_channel(cRaid.com, name=re.sub(r"-[0-9]*", str("-%i" %(len(cRaid.raid.participants))), cRaid.com.name))
+    await editListe(cRaid)
+    return 1
+
+#gestionnaires de la channel d'accueil
+async def addLevel(lvl, member):
+    """retourn 1 si l'ajout nu niveau a fonctionné
+    0 sinon"""
+    if not isinstance(lvl, int): return 0
+    if not isinstance(member, discord.Member): return 0
+
+    regex = re.compile(r"^.* \([0-9]*\)$") #un nick avec un niveau
+    if regex.match(member.nick):
+        newNick = re.sub(r"\([0-9]*\)$", str("(%i)" %(lvl)), member.nick)
+    else:
+        newNick = member.nick + str(" (%i)" %(lvl))
+
+    await client.change_nickname(member, newNick)
+async def changeTeam(team, user):
+    """enleve tous les rôles d'un utilisateur sauf @everyone et @modo puis place le member dans la team appropriée
+    return 1 si le changement est effectif 0 sinon"""
+    if not isinstance(member, discord.Member): return 0
+
+    for role in member.roles:
+        if not (role.name == "@everyone" or role.name == "@modo"): await client.delete_role(member, role)
+    await client.add_roles(member, next(r for r in server.roles if r.name == team))
 
 
-# timer toutes les minutes parcourir les raids
-    # si heure de fin dépassée alors on libere le salon
+# timer toutes les 10s
 async def waitTimer():
     while True:
         await asyncio.sleep(10)
-        print ("test")
 
         regex = re.compile(r"[0-9]*_[a-z0-9]*-[0-9]*") #nom des channels de raid
 
@@ -83,60 +107,52 @@ async def waitTimer():
                 cRaidCurrent = cRaids[numRaid]
                 date = datetime.datetime.now()
                 if cRaidCurrent.raid.fin < date:
-                    if not cRaidCurrent.retirerRaid():
-                        print("error retirer")
-                        continue
-                    if not await removeCRaid(cRaidCurrent):
-                        print("error remove ")
-                        continue
+                    if not cRaidCurrent.retirerRaid(): continue
+                    if not await removeCRaid(cRaidCurrent): continue
 
-
+#routine demarage
 @client.event
 async def on_ready():
     #variable externes
-    global activesChannels
     global cAccueil
-    global cRaidList
     global cDiscussion
     global cPokemon
     global cRaidAdd
     global server
-    global msgRaid
 
     #recuperer le server
     server = client.get_server(os.environ["DISCORD_SERVER_ID"])
 
     #on identifie tous les salon sur lesquel peut agir le bot
     regex = re.compile(r"[0-9]*_[a-z0-9]*-[0-9]*") #nom des channels de raid
-    cIds = []
+    cToDelete = []
     for cCurrent in client.get_all_channels():
         if cCurrent.name == "accueil":
             cAccueil = cCurrent
-        elif cCurrent.name == "raid-list":
-            cRaidList = cCurrent
         elif cCurrent.name == "discussion":
             cDiscussion = cCurrent
         elif cCurrent.name == "pokemon":
             cPokemon = cCurrent
-        elif cCurrent.name == "raid-add":
+        elif cCurrent.name == "raid":
             cRaidAdd = cCurrent
             await client.purge_from(cRaidAdd)
         elif regex.match(cCurrent.name):
-            cIds.append(cCurrent.id)
+            cToDelete.append(cCurrent)
 
-        for cId in cIds:
-            await client.delete_channel(client.get_channel(cId))
-    activeChannels = client.get_all_channels()
+        for cCurrent in cToDelete:
+            await client.delete_channel(cCurrent)
 
     #ecrire le message initiale des raid
-    msgRaid = await client.send_message(cRaidAdd, "liste des raides en cours")
+    await client.send_message(cRaidAdd, "liste des raides en cours")
 
     print("Bot is ready and back online !")
 
     #on lance le garbage collector
     await waitTimer()
 
+    #si on atteint cet endroit c'est que le garbage collector a craché
     print ('bug')
+
 #ajout manuel d'evenement
 @client.event
 async def on_message(message):
@@ -148,6 +164,8 @@ async def on_message(message):
     global cRaidList
     global cAccueil
     global server
+
+    #variables internes
     args = message.content.split(" ")
     regex = re.compile(r"[0-9]*_[a-z0-9]*-[0-9]*") #nom des channels de raid
     if message.content == "cookie":
@@ -155,6 +173,7 @@ async def on_message(message):
         await client.send_message(message.channel, "%i :cookie:" %(cookieCompteur) )
         await client.delete_message(message)
 
+    #on écoute la channel d'add
     elif message.channel == cRaidAdd:
         if message.content.lower().startswith("add") and not len(args) < 4:
             pokeName = args[1]
@@ -163,34 +182,23 @@ async def on_message(message):
 
             cCom = await client.create_channel(server, str("%i_%s-0" %(ChannelRaid.nb_channel+1,pokeName)))
             cRaids[ChannelRaid.nb_channel] = ChannelRaid(cCom)
-
             raid = Raid(0,pokeName,message.author, battleTime, battlePlace)
             cRaid = cRaids[ChannelRaid.nb_channel].ajouterRaid(raid)
 
-            await addToListe(cRaid)
-            msg = await client.send_message(cCom, embed=raid.embed())
-            await client.pin_message(msg)
-            cRaid.pinMsg = msg
+            if not await addToListe(cRaid): return
+            cRaid.pinMsg = await client.send_message(cCom, embed=raid.embed())
+            await client.pin_message(cRaid.pinMsg)
 
     #on écoute la channel d'accueil
     elif message.channel == cAccueil:
-        if message.content.startswith("lvl") and len(args) == 2:
-            lvl = args[1]
-            regex = re.compile(r"^.* \([0-9]*\)$")
-            print(message.author.nick)
-            if regex.match(message.author.nick):
-                newNick = re.sub(r"\([0-9]*\)$", str("(%s)" %(lvl)), message.author.nick)
-            else:
-                newNick = message.author.nick
-                newNick += str(" (%s)" %(lvl))
-            print(newNick)
-            await client.change_nickname(message.author, newNick)
-
-        if message.content.startswith("team") and len(args) == 2:
+        if message.content.lower().startswith("lvl") and len(args) == 2:
+            lvl = int(args[1])
+            if not await addLevel(lvl, message.author): return
+            await client.delete_message(message)
+        if message.content.lower().startswith("team") and len(args) == 2:
             team = args[1]
-            for role in message.author.roles:
-                if not role.name == "@everyone" : await client.delete_role(message.author, role)
-            await client.add_roles(message.author, next(r for r in server.roles if r.name == team))
+            if not await changeTeam(team, user): return
+            await client.delete_message(message)
 
 
     #écoute des channels de raid
@@ -199,58 +207,45 @@ async def on_message(message):
         cCurrent = cRaids[numRaid]
         if cCurrent.isRaid():
             if message.content.lower() == "in":
-                    if cCurrent.raid.ajouterParticipant(message.author):
-                        await client.edit_message(cCurrent.pinMsg, embed=cCurrent.raid.embed())
-                        await client.edit_channel(cCurrent.com, name=re.sub(r"-[0-9]*", str("-%i" %(len(cCurrent.raid.participants))), cCurrent.com.name))
-                        await editListe(cCurrent)
-                        await client.delete_message(message)
+                if not cCurrent.raid.ajouterParticipant(message.author): return
+                if not await editCRaid(cCurrent): return
+                await client.delete_message(message)
             elif message.content.lower() == "out":
-                if cCurrent.raid.retirerParticipant(message.author):
-                     await client.edit_message(cCurrent.pinMsg, embed=cCurrent.raid.embed())
-                     await client.edit_channel(cCurrent.com, name=re.sub(r"-[0-9]*", str("-%i" %(len(cCurrent.raid.participants))), cCurrent.com.name))
-                     await editListe(cCurrent)
-                     await client.delete_message(message)
+                if not cCurrent.raid.retirerParticipant(message.author): return
+                if not await editCRaid(cCurrent): return
+                await client.delete_message(message)
             elif message.content.lower() == 'abort':
-                if message.author == cCurrent.raid.capitaine:
-                    if cCurrent.retirerRaid():
-                        cId = cCurrent.com.id
-                        await removeFromListe(cCurrent)
-                        await client.delete_channel(client.get_channel(cId))
-                        cCurrent = 0
+                if not message.author == cCurrent.raid.capitaine: return
+                if not cCurrent.retirerRaid(): return
+                if not await removeCRaid(cCurrent): return
             elif args[0] == "launch" and len(args) == 2:
                 battleTime = args[1]
-                if cCurrent.raid.choisirLaunch(battleTime):
-                    await client.edit_message(cCurrent.pinMsg, embed=cCurrent.raid.embed())
-                    await editListe(cCurrent)
-                    await client.delete_message(message)
+                if not cCurrent.raid.choisirLaunch(battleTime): return
+                if not await editCRaid(cCurrent): return
+                await client.delete_message(message)
             elif args[0].lower() == "edit" and len(args) == 2:
                 pokeName = args[1]
-                if cCurrent.raid.faireEclore(pokeName):
-                    await client.edit_message(cCurrent.pinMsg, embed=cCurrent.raid.embed())
-                    await client.edit_channel(cCurrent.com, name=re.sub(r"_[a-z0-9]*-", str("_%s-" %(pokedex[cCurrent.raid.pokeId-1]["fr"])), cCurrent.com.name))
-                    await editListe(cCurrent)
-                    await client.delete_message(message)
+                if not cCurrent.raid.faireEclore(pokeName): return
+                if not await editCRaid(cCurrent): return
+                await client.delete_message(message)
             elif args[0].lower() == "dispo" and len(args) == 2:
                 userId = args[1].replace('<@', '').replace('>', '')
-                if cCurrent.raid.ajouterParticipant(next( m for m in client.get_all_members() if m.id == userId)):
-                    await client.edit_message(cCurrent.pinMsg, embed=cCurrent.raid.embed())
-                    await client.edit_channel(cCurrent.com, name=re.sub(r"-[0-9]*", str("-%i" %(len(cCurrent.raid.participants))), cCurrent.com.name))
-                    await editListe(cCurrent)
+                if not cCurrent.raid.ajouterParticipant(next( m for m in client.get_all_members() if m.id == userId)): return
+                if not await editCRaid(cCurrent): return
 
+#ajout d'emoji
 @client.event
 async def on_reaction_add(reaction, user):
-
     regex = re.compile(r"[0-9]*_[a-z0-9]*-[0-9]*") #nom des channels de raid
     if regex.match(reaction.message.channel.name):
         numRaid = int(reaction.message.channel.name[0])
         cCurrent = cRaids[numRaid]
         if cCurrent.isRaid():
             if reaction.emoji == '👌':
-                if cCurrent.raid.ajouterParticipant(user):
-                    await client.edit_message(cCurrent.pinMsg, embed=cCurrent.raid.embed())
-                    await client.edit_channel(cCurrent.com, name=re.sub(r"-[0-9]*", str("-%i" %(len(cCurrent.raid.participants))), cCurrent.com.name))
-                    await editListe(cCurrent)
+                if not cCurrent.raid.ajouterParticipant(user): return
+                if not await editCRaid(cCurrent): return
 
+#retrait d'emoji
 @client.event
 async def on_reaction_remove(reaction, user):
     regex = re.compile(r"[0-9]*_[a-z0-9]*-[0-9]*") #nom des channels de raid
@@ -259,9 +254,7 @@ async def on_reaction_remove(reaction, user):
         cCurrent = cRaids[numRaid]
         if cCurrent.isRaid():
             if reaction.emoji == '👌':
-                if cCurrent.raid.retirerParticipant(user):
-                    await client.edit_message(cCurrent.pinMsg, embed=cCurrent.raid.embed())
-                    await client.edit_channel(cCurrent.com, name=re.sub(r"-[0-9]*", str("-%i" %(len(cCurrent.raid.participants))), cCurrent.com.name))
-                    await editListe(cCurrent)
+                if not cCurrent.raid.retirerParticipant(user): return
+                if not await editCRaid(cCurrent): return
 
 client.run(os.environ['DISCORD_TOKEN'])
